requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
audio-recorder-streamlit==0.0.8
Babel
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.2
bleach
blinker==1.8.2
blis==0.7.11
boto3==1.34.90
botocore==1.34.90
Brotli
cached-property
cachetools==5.3.3
catalogue==2.0.10
certifi
cffi
chardet==5.2.0
charset-normalizer
click==8.1.7
cloudpathlib==0.16.0
cohere==5.3.4
colorama
comm
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.5
debugpy
decorator
deepdiff==7.0.1
deeplake==3.9.4
defusedxml
dill==0.3.8
distro==1.9.0
docopt==0.6.2
elevenlabs==0.2.18
emoji==2.11.1
en-core-web-sm
entrypoints
exceptiongroup
executing
faiss-cpu==1.8.0
fastavro==1.9.4
fastjsonschema
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.4
filetype==1.2.0
fqdn
frozenlist==1.4.1
fsspec==2024.3.1
gguf==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11
h2
hpack==4.0.0
httpcore
httplib2==0.22.0
httpx
httpx-sse==0.4.0
huggingface-hub==0.22.2
humbug==0.3.2
hyperframe
idna
importlib-metadata==6.11.0
importlib_resources
intel-openmp==2021.4.0
ipykernel
ipython==8.12.3
isoduration
jaraco.context==5.3.0
jedi
jieba3k==0.35.1
Jinja2
jmespath==1.0.1
joblib==1.4.0
json5
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.0.4
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainplus-sdk==0.0.20
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.54
language_data==1.2.0
llvmlite==0.42.0
lxml==5.2.1
lxml_html_clean==0.1.1
lz4==4.3.3
marisa-trie==1.1.0
markdown-it-py==3.0.0
MarkupSafe
marshmallow==3.21.1
marshmallow-enum==1.5.1
matplotlib-inline
mdurl==0.1.2
mistune
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
nbclient
nbconvert
nbformat
nest_asyncio
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
notebook_shim
numba==0.59.1
numcodecs==0.12.1
numexpr==2.10.0
numpy==1.26.4
openai==1.26.0
openai-whisper
openapi-schema-pydantic==1.2.4
ordered-set==4.1.0
orjson==3.10.3
outcome==1.3.0.post0
overrides
packaging==23.2
pandas==2.2.2
pandocfilters
parso
pathos==0.3.2
pickleshare
Pillow==9.5.0
pipreqs==0.5.0
pkgutil_resolve_name
platformdirs
pox==0.3.4
ppft==1.7.6.8
preshed==3.0.9
prometheus_client
prompt-toolkit
proto-plus==1.23.0
protobuf==4.25.3
psutil
pure-eval
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser
pycryptodomex==3.20.0
pydantic==1.10.15
pydeck==0.9.0
Pygments
PyJWT==2.8.0
pyllamacpp==2.4.3
Pympler==1.0.1
pyparsing==3.1.2
pypdf==4.2.0
PySocks
python-dateutil
python-dotenv
python-iso639==2024.4.27
python-json-logger
python-magic==0.4.27
pytz
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
pywin32==306
pywinpty
PyYAML
pyzmq
rapidfuzz==3.8.1
referencing
regex==2024.4.16
requests
requests-file==2.0.0
rfc3339-validator
rfc3986-validator
rich==13.7.1
rpds-py
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.15.2
Send2Trash
sentencepiece==0.2.0
setuptools==69.5.1
sgmllib3k==1.0.0
six
smart-open==6.4.0
smmap==5.0.1
sniffio
sortedcontainers==2.4.0
soupsieve
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data
streamlit==1.23.1
streamlit-chat==0.0.2.2
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.3
terminado
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tinycss2
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli
toolz==0.12.1
torch==2.3.0
tornado
tqdm==4.66.2
traitlets
transformers==4.40.1
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
types-python-dateutil
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing-utils
typing_extensions
tzdata==2024.1
tzlocal==4.3.1
unstructured==0.11.8
unstructured-client==0.22.0
uri-template
uritemplate==4.1.1
urllib3
validators==0.28.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth
weasel==0.3.4
webcolors
webencodings
websocket-client
websockets==12.0
wheel==0.43.0
wikipedia==1.4.0
win-inet-pton
wolframalpha==5.0.0
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
yarg==0.1.9
yarl==1.9.4
yt-dlp==2024.4.9
zipp